This is acceptable to png_set_IHDR().  Now to make sure it stays that way.
In the old style, the individual parts of the header were set up as they
were figured out.  In the new style, this must be done when everything
is figured out -- with a call to png_set_IHDR().  So, I will be
progressing down the body of WritePNG(), feeding dummy parameters to
png_set_IHDR() until all parameters are determined.  Then the call to
png_set_IHDR() will settle down near the end of WritePNG() and this
problem with "bummer!" errors will be fixed.