Use correct tick definition of the system instead of a hard-defined value.
This is from xv-20091215-werner-fink-multiple-fixes.patch.mime.
CLK_TCK isn't defined anymore on newer Linux systems as it now becomes
variable at boot time.  I need to revisit this after this patch-a-thon
to make sure that if CLK_TCK is indeed defined, then we should use that
so as to keep backwards compatibility.