Fix off-by-one problem causing copy-and-paste corruption when in 24-bit mode.
This is from xv-20100523-ville-saari-clipboard-colormap-fix.dif

The report states that it happens when the selected area has 256
distinct colors or fewer.

Here's how to trigger the bug:  To make it easy, I picked a .gif with
just three colors: blue (000084), red (ff0000), and white (ffffff).
Well...  Three colors plus the other ones in between to do
anti-aliasing.  Anyhow...

Press Meta-8 (Alt-8) to go into 24-bit mode. Select a small area, maybe
60 pixels wide.  Crop it (press 'c'), then stretch out the image window
so the individual pixels are around 5mm on your monitor.  Select a piece
of image.  Right-click on the image to get the controls window.  Press
the "copy" button.  Click the image outside the selection box to get rid
of the box.  Press the "paste" button.  This will bring up a new
selection box which indicates where the contents of the clipboard will
be pasted.  Move that box somewhere else.  Press the "paste" button
again.  What you paste will be sprinkled with what looks like confetti.