Eliminate flicker
This is from xv-michal-maruska-no-flicker-bkgd.patch.

Hello, here's a patch for XV:

Since the Image window is full of the image pixels, the window background
is not meant to be seen.... but it can make flicker, be seen for an instant,
when the window is exposed.

More info at:         http://michal.ruska.dyndns.org/flicker.html